Fixed jars being unpacked every time and when you are moving around the mouse it will jump to the other side of the screen

+++ src/com/sekwah/modeleditor/files/Unpacker.java
@@ -24,34 +24,38 @@ public class Unpacker implements Runnable{
 			JarFile jar = new JarFile(jarFile);
 			Enumeration enum1 = jar.entries();
 			float totalDataRead = 0;
-			while(enum1.hasMoreElements())
-			{
-				try{
-				float percent = (++totalDataRead/jar.size())*100;//fix
-				JarEntry file = (JarEntry)enum1.nextElement();
-				if(!file.isDirectory()){
-				File f = new File(File.separator + destDir + File.separator + file.getName());
-				System.out.println("Unpacking: " + f.getAbsolutePath() + "(" + file.getName() + ")");
-				f.getParentFile().mkdirs();
-				InputStream is = jar.getInputStream(file);
-				FileOutputStream fos = new FileOutputStream(f);
-				
-				byte[] buf = new byte[4096];
-			    int r;
-			    while ((r = is.read(buf)) != -1) {
-			      fos.write(buf, 0, r);
-			    }
-			    
-				/**while(is.available() > 0)
-				{
-					//ProgressBar.setValue(100 - (int) (((float)is.available() / (float)startAvailable) * 100));
-					fos.write(is.read());
-				}*/
-				fos.close();
-				is.close();
+			while(enum1.hasMoreElements()) {
+				try {
+					float percent = (++totalDataRead / jar.size()) * 100;//fix
+					JarEntry file = (JarEntry) enum1.nextElement();
+					if (!file.isDirectory()) {
+						File f = new File(File.separator + destDir + File.separator + file.getName());
+						if (f.exists()) {
+							System.out.println("Found: " + f.getAbsolutePath() + "(" + file.getName() + ")");
+							continue;
+						}
+						System.out.println("Unpacking: " + f.getAbsolutePath() + "(" + file.getName() + ")");
+						f.getParentFile().mkdirs();
+						InputStream is = jar.getInputStream(file);
+						FileOutputStream fos = new FileOutputStream(f);
+
+						byte[] buf = new byte[4096];
+						int r;
+						while ((r = is.read(buf)) != -1) {
+							fos.write(buf, 0, r);
+						}
+
+						/**while(is.available() > 0)
+						 {
+						 //ProgressBar.setValue(100 - (int) (((float)is.available() / (float)startAvailable) * 100));
+						 fos.write(is.read());
+						 }*/
+						fos.close();
+						is.close();
+					}
+				} catch (Exception e) {
+					e.printStackTrace();
 				}
-				}
-				catch(Exception e){e.printStackTrace();}
 			}
 		}
 		catch(Exception e){e.printStackTrace();}
